+++ front-end/public/index.html
@@ -11,7 +11,7 @@
     <noscript>
       <strong>We're sorry but <%= htmlWebpackPlugin.options.title %> doesn't work properly without JavaScript enabled. Please enable it to continue.</strong>
     </noscript>
-    <div id="app"></div>
+    <div id="dashboard"></div>
     <!-- built files will be auto injected -->
   </body>
 </html>
